+++ cudaApplication/src/utils/matrix_util.cu
@@ -1,13 +1,13 @@
 #include "matrix_util.cuh"
 
-std::vector<Triangle> rotate_mesh(Mesh& mesh, double rX, double rY, double rZ){
-    Eigen::Matrix3d rotMatX;
-    rotMatX = Eigen::AngleAxisd(rX, Eigen::Vector3d::UnitX());
-    Eigen::Matrix3d rotMatY;
-    rotMatY = Eigen::AngleAxisd(rY, Eigen::Vector3d::UnitY());
-    Eigen::Matrix3d rotMatZ;
-    rotMatZ = Eigen::AngleAxisd(rZ, Eigen::Vector3d::UnitZ());
-    Eigen::Matrix3d rotationMatrix = rotMatZ * rotMatY * rotMatX;
+std::vector<Triangle> rotate_mesh(Mesh& mesh, float rX, float rY, float rZ){
+    Eigen::Matrix3f rotMatX;
+    rotMatX = Eigen::AngleAxisd(rX, Eigen::Vector3f::UnitX());
+    Eigen::Matrix3f rotMatY;
+    rotMatY = Eigen::AngleAxisd(rY, Eigen::Vector3f::UnitY());
+    Eigen::Matrix3f rotMatZ;
+    rotMatZ = Eigen::AngleAxisd(rZ, Eigen::Vector3f::UnitZ());
+    Eigen::Matrix3f rotationMatrix = rotMatZ * rotMatY * rotMatX;
     std::vector<Triangle> rotated_triangles = mesh.triangles;
     for (auto& tri : rotated_triangles) {
         tri.p1 = rotationMatrix * tri.p1;
@@ -17,8 +17,8 @@ std::vector<Triangle> rotate_mesh(Mesh& mesh, double rX, double rY, double rZ){
     return rotated_triangles;
 }
 
-std::vector<Triangle> translate_mesh(Mesh& mesh, double tx, double ty, double tz) {
-    Eigen::Vector3d translation(tx, ty, tz);
+std::vector<Triangle> translate_mesh(Mesh& mesh, float tx, float ty, float tz) {
+    Eigen::Vector3f translation(tx, ty, tz);
     std::vector<Triangle> translated_triangles = mesh.triangles;
     for (auto& tri : translated_triangles) {
         tri.p1 += translation;
